requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dataclasses-json==0.6.7
EbookLib==0.19
frozenlist==1.7.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
IceSpringPySideStubs-PyQt6==1.3.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-ollama==0.3.8
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langsmith==0.4.31
lxml==6.0.2
markdown2==2.5.4
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
ollama==0.6.0
orjson==3.11.3
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
striprtf==0.0.29
tenacity==9.1.2
tiktoken==0.11.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
xlsxwriter==3.2.9
yarl==1.20.1
zstandard==0.25.0
